improve metric values calculation, add kvm_cpuid parser
metrics calculation:
-treat the metric value as float
	before it was int
-add input parameter time for rate calculation
	before it produced cumulative sum only
-spaces removal
	some code cleaning

kvm_cpuid parser:
the parser extracts EAX values used to execute cpuid instruction

+++ kvm_trace_histogram/show_histogram.py
@@ -10,7 +10,7 @@ from matplotlib.lines import Line2D
 
 def get_points(points):
 	names = [point["name"] for point in points]
-	vals = [int(point["val"]) for point in points]
+	vals = [float(point["val"]) for point in points]
 	r = dict()
 	r["names"] = names
 	r["vals"] = vals
